Disable 'Show hidden' button when no content is hidden.
This commit implements the functionality to disable the 'Show hidden' button
when there are no hidden posts or threads on the page. The button's tooltip
will also dynamically update to reflect whether there are hidden posts or
threads, providing clear feedback to the user. This improves the user
experience by preventing unnecessary clicks and providing relevant information.